requirements.txt
anthropic>=0.25.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
flask>=3.0.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
python-telegram-bot>=21.0
requests>=2.31.0
py-clob-client>=0.14.0
eth-account>=0.10.0
gunicorn>=21.2.0
